requirements.txt
antlr4-python3-runtime==4.11.1
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.8
astunparse==1.6.3
attrs==20.3.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
## !! Could not determine repository location
-e /Users/arfat/Desktop/bloqs
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cirq==1.1.0
cirq-aqt==1.1.0
cirq-core==1.1.0
cirq-google==1.1.0
cirq-ionq==1.1.0
cirq-pasqal==1.1.0
cirq-rigetti==1.1.0
cirq-web==1.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
coverage==6.5.0
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
duet==0.2.7
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
fonttools==4.37.4
fuzzywuzzy==0.18.0
google-api-core==1.33.1
google-auth==2.12.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
grpcio==1.49.1
grpcio-status==1.48.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
iso8601==1.1.0
jedi==0.18.1
Jinja2==3.1.2
jmetalpy==1.5.5
joblib==1.2.0
json5==0.9.10
jsonpath-ng==1.5.3
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.2
kiwisolver==1.4.4
lark==0.11.3
lark-parser==0.12.0
Levenshtein==0.20.9
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
msgpack==0.6.2
mypy-extensions==0.4.3
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.4.12
notebook-shim==0.1.0
ntlm-auth==1.5.0
numpy==1.23.3
openqasm3==0.4.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
patsy==0.5.2
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.11.0
ply==3.11
prettytable==3.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyGithub==1.58.0
Pygments==2.13.0
PyJWT==2.5.0
pylatexenc==2.10
PyNaCl==1.5.0
pyparsing==3.0.9
pyquil==3.3.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python-rapidjson==1.8
pytket==1.11.1
pytket-cirq==0.26.0
pytket-qiskit==0.28.0
pytket-qsharp==0.32.0
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qcs-api-client==0.21.2
qiskit==0.41.0
qiskit-aer==0.11.2
qiskit-ibmq-provider==0.20.0
qiskit-qasm3-import==0.1.0
qiskit-terra==0.23.2
qsharp==0.27.253010
qsharp-chemistry==0.27.253010
qsharp-core==0.27.253010
qwasm==1.0.1
rapidfuzz==2.13.7
requests==2.28.1
requests-ntlm==1.1.0
retry==0.9.2
retrying==1.3.3
retworkx==0.11.0
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.10.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
rustworkx==0.12.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
statsmodels==0.13.2
stevedore==4.0.0
symengine==0.9.2
sympy==1.11.1
tenacity==8.1.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
tweedledum==1.1.1
types-pkg-resources==0.1.3
types-python-dateutil==2.8.19
types-retry==0.9.9
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
wrapt==1.14.1
zipp==3.8.1
